requirements.txt
# Core ML and Data Processing
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0

# Deep Learning
tensorflow>=2.4.0
keras>=2.4.0
h5py>=3.1.0

# Image Processing
Pillow>=8.0.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.3.2
seaborn>=0.11.0

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Utilities
tqdm>=4.50.0
python-dotenv>=0.19.0
requests>=2.26.0

# Testing and Development
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.0